riscv: cheri: Fix arch/riscv/kernel/stacktrace.c
Properly cast between uintptr_t and real pointers. This implies
changes to related header files.

+++ include/linux/sched/debug.h
@@ -2,6 +2,8 @@
 #ifndef _LINUX_SCHED_DEBUG_H
 #define _LINUX_SCHED_DEBUG_H
 
+#include <linux/types.h>
+
 /*
  * Various scheduler/task debugging interfaces:
  */
@@ -30,7 +32,7 @@ extern void show_regs(struct pt_regs *);
  * task), SP is the stack pointer of the first frame that should be shown in the back
  * trace (or NULL if the entire call-chain of the task should be shown).
  */
-extern void show_stack(struct task_struct *task, unsigned long *sp,
+extern void show_stack(struct task_struct *task, uintptr_t *sp,
 		       const char *loglvl);
 
 extern void sched_show_task(struct task_struct *p);
